enrichment-agent/scalability_requirements.txt
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0 